+++ src/config/level/symfony/symfony-constructor-injection.yml
@@ -1,0 +1,3 @@
+rectors:
+    Rector\Rector\Contrib\Symfony\FrameworkBundle\ContainerGetToConstructorInjectionRector: ~
+    Rector\Rector\Contrib\Symfony\FrameworkBundle\GetParameterToConstructorInjectionRector: ~
