+++ tests/Rector/Contrib/Symfony/FrameworkBundle/GetToConstructorInjectionRector/Correct/correct.php.inc
