+++ tests/Rector/Contrib/Symfony/FrameworkBundle/GetToConstructorInjectionRector/GetToConstructorInjectionRectorTest.php
@@ -1,13 +1,13 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector;
+namespace Rector\Tests\Rector\Contrib\Symfony\FrameworkBundle\GetToConstructorInjectionRector;
 
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
 /**
- * @covers \Rector\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector
+ * @covers \Rector\Rector\Contrib\Symfony\FrameworkBundle\GetToConstructorInjectionRector
  */
-final class GetterToPropertyRectorTest extends AbstractRectorTestCase
+final class GetToConstructorInjectionRectorTest extends AbstractRectorTestCase
 {
     /**
      * @dataProvider provideWrongToFixedFiles()
@@ -29,6 +29,6 @@ final class GetterToPropertyRectorTest extends AbstractRectorTestCase
 
     protected function provideConfig(): string
     {
-        return __DIR__ . '/Source/rector.yml';
+        return __DIR__ . '/config.yml';
     }
 }
