+++ tests/Rector/Contrib/Symfony/FrameworkBundle/GetToConstructorInjectionRector/config.yml
@@ -1,0 +1,9 @@
+rectors:
+    Rector\Rector\Contrib\Symfony\FrameworkBundle\GetToConstructorInjectionRector: ~
+
+services:
+    Rector\Tests\Rector\Contrib\Symfony\FrameworkBundle\GetToConstructorInjectionRector\Source\DummyProvider: ~
+
+    # require for interface autowiring
+    'Rector\Contract\Bridge\ServiceTypeForNameProviderInterface':
+        alias: 'Rector\Tests\Rector\Contrib\Symfony\FrameworkBundle\GetToConstructorInjectionRector\Source\DummyProvider'
