--- tests/Rector/Contrib/Symfony/HttpKernel/GetterToPropertyRector/Source/rector.yml
@@ -1,8 +1,0 @@
-rectors:
-    Rector\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector: ~
-
-services:
-    Rector\Tests\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector\Source\DummyProvider: ~
-
-    # require for interface autowiring
-    'Rector\Contract\Bridge\ServiceTypeForNameProviderInterface': '@Rector\Tests\Rector\Contrib\Symfony\HttpKernel\GetterToPropertyRector\Source\DummyProvider'
